requirements.txt
requests
python-daemon
pyfiglet
rich
pytest
pyyaml
